+++ XenAdmin/Dialogs/VMDialogs/SelectVMsToSuspendDialog.cs
@@ -101,7 +101,7 @@ namespace XenAdmin.Dialogs.VMDialogs
                 if (row.Action != ActionCellAction.Migrate)
                     continue;
 
-                required += row.Vm.memory_dynamic_max;
+                required += (long)row.Vm.memory_dynamic_max;
             }
             RequiredLabel.ForeColor = required > PoolMemoryFree ? Color.Red : ForeColor;
             RequiredLabel.Text = Util.MemorySizeStringMB(required);
